Move variable initializations out of constructor into declarations and eliminate goto statements. Rename adis16448.cpp/h to ADIS16448.cpp/h. Added ADIS16448_accel.cpp/h, ADIS16448_gyro.cpp/h, ADIS16448_mag.cpp/h, and added ADIS16448_accel class to standardize inheriting classes.
