+++ src/drivers/drv_accel.h
@@ -52,12 +52,12 @@
 
 /** accel scaling factors; Vout = Vscale * (Vin + Voffset) */
 struct accel_calibration_s {
-	float	x_offset;
-	float	x_scale;
-	float	y_offset;
-	float	y_scale;
-	float	z_offset;
-	float	z_scale;
+	float x_offset;
+	float y_offset;
+	float z_offset;
+	float x_scale;
+	float y_scale;
+	float z_scale;
 };
 /*
  * ioctl() definitions
